Implement delegation architecture for inter-app communication
NOTES: NotificationCenter design was trashed because not all views needed to be notified of the same notifications. Instead, multiple delegates are created: (1) delegate for the main menu, (2) delegate for the scanner, (3) and delegate for any of the motor based views (i.e. panorama and manual control). The first only needs to know if the app is connected to a device. The second only needs to know if Bluetooth is turned on or not. The third only needs to know if the device connection has been interrupted. (An optional reconnect delegate may be implemented as well.)